requirements.txt
ollama
typing
streamlit
python-docx
logging
langchain
langchain_community
pytest
SpeechRecognition
pydub
moviepy
PyPDF2
pytube
reportlab
langdetect
deep-translator
websockets
asyncio
json-diff
